refactor: extract KeycapSymbols utility from OverlayKeycapView extensions
Move pure rendering utilities (sfSymbolForAction, isModifierOrSpecialKey,
dynamicTextLabel, windowActionSymbol, windowActionColor, collectionColor)
into a standalone KeycapSymbols enum. Extension files become thin
forwarding wrappers. No rendering changes — snapshot tests pass
pixel-identical.

Phase 4.1 of the overlay keycap decomposition plan.
